Rename a folder to get a better icon
